+++ tests/test_mapper_basic.py
@@ -15,7 +15,7 @@ class TestMapperBasic(unittest.TestCase):
         user = mock_galaxy.User('gargravarr', 'fairycake@vortex.org')
         tpv_config = tpv_config_path or os.path.join(os.path.dirname(__file__),
                                                      'fixtures/mapping-basic.yml')
-        gateway.ACTIVE_DESTINATION_MAPPER = None
+        gateway.ACTIVE_DESTINATION_MAPPERS = {}
         return gateway.map_tool_to_destination(galaxy_app, job, tool, user, tpv_config_files=[tpv_config])
 
     def test_map_default_tool(self):