+++ tpv/core/loader.py
@@ -4,12 +4,16 @@ import ast
 import functools
 import logging
 from typing import Dict
+import re
 
 from . import helpers, util
 from .entities import Entity, GlobalConfig, TPVConfig
 from .evaluator import TPVCodeEvaluator
 
 log = logging.getLogger(__name__)
+
+
+NOQA_RE = re.compile(r"#\s*noqa:\s*([A-Z0-9, ]+)?")
 
 
 class InvalidParentException(Exception):
@@ -88,6 +92,27 @@ class TPVConfigLoader(TPVCodeEvaluator):
     def validate_entities(self, entities: Dict[str, Entity]) -> dict:
         self.recompute_inheritance(entities)
 
+    def get_noqa_codes(self, entity_comments: list) -> (bool, set[str] | None):
+        comments = []
+        if entity_comments and len(entity_comments) == 4 and entity_comments[3]:
+            comments.extend([x.value.strip() for x in entity_comments[3]])
+
+        for comment in comments:
+            match = re.match(r"#\s*noqa:?\s*([A-Z0-9, ]+)?", comment)
+            if match:
+                codes = match.group(1)
+                # Return a set of codes or None if `# noqa` with no codes
+                return (True, set(code.strip() for code in codes.split(',')) if codes else None)
+
+        return (False, None)
+
+    def store_noqa_codes(self, entity_list: dict, entity_id: str, noqa_dict: dict):
+        if hasattr(entity_list, "ca"):
+            entity_comments = entity_list.ca.items.get(entity_id)
+            noqa, noqa_codes = self.get_noqa_codes(entity_comments)
+            if noqa:
+                noqa_dict[entity_id] = noqa_codes
+
     def process_entities(self, tpv_config: TPVConfig) -> dict:
         self.validate_entities(tpv_config.tools),
         self.validate_entities(tpv_config.users),
@@ -124,6 +149,21 @@ class TPVConfigLoader(TPVCodeEvaluator):
 
     def merge_loader(self, loader: TPVConfigLoader):
         self.merge_config(loader.config)
+
+    def check_noqa(self, entity: Entity, code: str) -> bool:
+        if type(entity) is Tool:
+            noqa = self.noqa['tools']
+        elif type(entity) is User:
+            noqa = self.noqa['users']
+        elif type(entity) is Role:
+            noqa = self.noqa['roles']
+        elif type(entity) is Destination:
+            noqa = self.noqa['destinations']
+        else:
+            raise RuntimeError(f"Unknown entity type: {entity}")
+        if entity.id in noqa and (noqa[entity.id] is None or code in noqa[entity.id]):
+            return True
+        return False
 
     @staticmethod
     def from_url_or_path(url_or_path: str):