+++ tpv/core/util.py
@@ -5,7 +5,7 @@ import requests
 
 
 def load_yaml_from_url_or_path(url_or_path: str):
-    yaml = ruamel.yaml.YAML(typ='safe')
+    yaml = ruamel.yaml.YAML(typ="rt")
     if os.path.isfile(url_or_path):
         with open(url_or_path, 'r') as f:
             return yaml.load(f)